jjb: Replace open_enoent testcase for a more stable testcase
The open_enoent testcase showed a bimodal distribution of time per event
that was measurable even when tracing was disabled. This behaviour was
witnessed only in cases where the open failed with an `enoent` error.
Our investigation of the sys_open code pointed towards to the pathwalk
in the VFS. We decided to simply change the test case to something more
stable due to lack of time to put in the investigation.

Signed-off-by: Francis Deslauriers <francis.deslauriers@efficios.com>