jjb: Remove internal Lava tests from the passing requirements
Lava runs a number of basic tests when launching a test node and some of
these tests would randomly be marked as failed without affecting the
behaviour of the actual tests or benchmarks.
To avoid a noisy CI, we simply omit to consider those tests when
checking for run success or failure.

Lava devs on IRC suggested to upgrade to the new Lava version and see if the
error was still present.

Signed-off-by: Francis Deslauriers <francis.deslauriers@efficios.com>