jjb: Keep the log of the last 100 `build_kernel_PARAM` jobs
When multiple commits are pushed on kernel and lttng-modules branches
close in time, a large number of kernel build jobs are launched.

The logs of build failures sometimes got deleted to make room for new
logs thus making the investigation for those failures impossible.

To prevent that, double the number of build logs kept.

Signed-off-by: Francis Deslauriers <francis.deslauriers@efficios.com>