jjb: lava: remove useless and periodically failing git clone
This git clone periodically failing with the following error:
```
Checking connectivity... fatal: missing blob object 'f805e810e5c6e087791506b4e721958de3574ae4'
fatal: remote did not send all necessary objects
Unexpected end of command stream
```
The main hypothesis for this failure is that it is caused because the
home directory of the use is an NFS mount.
Eventhough the root cause is believed the be found. It was later found
that this git clone is useless since the lava benchmark jobs are taking
care of checking out the Github repo.

Signed-off-by: Francis Deslauriers <francis.deslauriers@efficios.com>