jjb: lava: Rework baremetal jobs and introduce trigger job
Previously, each job was tracking jobs in its branches of interest in an
odd combination Jenkins CMS trigger and Bash script. It readability was
low and maintainability even lower.

This commit implements those features using the Groovy script console.

Lava jobs are now taking commit IDs as parameter. The trigger job checks
the status of each git branches of interest and launch the right jobs
for each new changes in those branches using the right commit IDs.

The goal of these changes is to increase maintainability and to regroup
the lava jobs under a single umbrella Jenkins project.

Signed-off-by: Francis Deslauriers <francis.deslauriers@efficios.com>