jjb: lava: kprobe-fuzzing: Don't fuzz ftrace functions
Probing ftrace functions like ftrace_ops_assist_func and some other
function can cause crashes and it won't be fixed by the kernel community
in the foreseeable future. See discussion [1].

People seem to agree it's a bug but we don't have a timeline on the fix.

For now, we ourselves manually blacklist the symbols from the fuzzed
instrumentation points.

[1] https://lkml.org/lkml/2018/3/16/560

Signed-off-by: Francis Deslauriers <francis.deslauriers@efficios.com>