LAVA: use xenial based rootfs instead of bionic
Use xenial instead of bionic ++ since syscall tests depends
on cat and the libc to use the open syscall. In recent libc openat is
used. See these commit in lttng-tools that helps with the problem:
    c8e51d1559c48a12f18053997bbcff0c162691c4
    192bd8fb712659b9204549f29d9a54dc2c57a9e

These are only part of 2.11 and were not backported since they do not
represent a *problem* per se.

Signed-off-by: Jonathan Rajotte <jonathan.rajotte-julien@efficios.com>