hang_processes.sh: Fix some ShellCheck warnings
Specifically:

In hang_processes.sh line 29:
if [ ! -z "$lttng_processes" ]; then
     ^-- SC2236: Use -n instead of ! -z.

In hang_processes.sh line 54:
if [ ! -z "$core_files" ]; then
     ^-- SC2236: Use -n instead of ! -z.

In hang_processes.sh line 67:
    rm -rf $core_files
           ^---------^ SC2086: Double quote to prevent globbing and word splitting.

I also removed the -r on the rm.  Since we are not deleting directories (and it
seems safer not to use -rf if we don't need to, in case $core_files contains
the wrong thing).

Signed-off-by: Simon Marchi <simon.marchi@efficios.com>