Fix: jjb: system-tests: pass modules repo to build job
When passing a custom Git repository URL to a systems-test job, the
master job doesn't pass this URL to the kernel and modules building job.
This makes the job fail when it can't checkout the desired commit on the
default remote.

To fix this, pass the URL received in the job parameter to the kernel
and modules building job.

Signed-off-by: Francis Deslauriers <francis.deslauriers@efficios.com>