Introduce Depends-on for lttng-tools gerrit job
The development of a lttng-tools feature can happen in parallel with a
lttng-ust feature. The need to build and test a lttng-tools gerrit
change against a gerrit lttng-ust change arises.

This patch introduce gerrit-based inter project dependency
testing for the lttng-tools project against lttng-ust.

The dependency can be expressed at the change's commit message level:

    Depends-on: lttng-ust: <lttng-ust-change-id>

e.g
    Fix: client_flush_outgoing_queue expect an active client

    Move validated and active assignation before the call to
    client_flush_outgoing_queue.

    Depends-on: lttng-ust: If8557b77f2a03b05b924cdc26ecf7618fff46289

    Signed-off-by: Jonathan Rajotte <jonathan.rajotte-julien@efficios.com>
    Change-Id: I873b2fad567eb6d59a53a12cba02451afd1d9827

For now, only lttng-ust is supported but babeltrace and lttng-modules
could someday be prime candidate for "Depends-on".

Note that the "Depends-on" line is optional. If none is present, we
import lttng-ust from the lttng-ust base job as it is done today.

For building ust, we reuse the script "scripts/lttng-ust/build.sh".

Note that we override the "$build" variable to "std" for the dependency
since the build type for the lttng-ust dependency is irrelevant for the
current job.

Signed-off-by: Jonathan Rajotte <jonathan.rajotte-julien@efficios.com>