jjb: lava: conditionally turn on `--enable-test-sdt-uprobe`
This flag is explicitly turned on so to fail if the feature dependencies
(`sdt.h` and `dtrace(1)`) are missing. This is to avoid the case where
the feature and its tests are silently turned off when these are
missing.

This flag is turned on when the LTTng version being tested is "master"
or "stable-2.11", and "stable-2.12" because only those versions have
this configure flag.

Signed-off-by: Francis Deslauriers <francis.deslauriers@efficios.com>