gerrit: depends-on: use master when dependency changes are merged
When the change a lttng-tools change depends on is merged/abandoned use
master as the ref.

This is not ideal CI time wise since we do not reuse artifact
but this solve a tricky situation were we actually want to use master
instead of the change available on gerrit. Intermediary changes
present in master could have an impact on the change under test.

Signed-off-by: Jonathan Rajotte <jonathan.rajotte-julien@efficios.com>