system-tests: improve lava slave ssh capability
The current lava slave setup does not permit easy access to the running
slave once the job is started. This can be cumbersome when a hang
happens. To ease access, setup ssh with authorized keys for the root
user. This is only accessible from internal network.

Split the env setup in two: base and vlttng.

Signed-off-by: Jonathan Rajotte <jonathan.rajotte-julien@efficios.com>