lttng-tools: executable cannot traverse root directory
The test_ust_notifier_discarded_count_multi_uid tests from lttng-tools
required that an instrumented application launched under a temporary
user be able to access lttng-ust libs in DEPS_LIBS. Since the dependancy
are not installed globally and that root owned directory are not
traversable by default, the apps fail to find liblttng-ust due to
permission setting.

To alleviate this for our CI, simply set the a+x bit on all leading
directly to DEPS_LIBS. This is acceptable under our CI since and the
root jobs since we always reset the work state to a known good state.

Signed-off-by: Jonathan Rajotte <jonathan.rajotte-julien@efficios.com>