Revert "jjb: tools: mark build as aborted on timeout."
This reverts commit 3b9bb48d192c642d57f678220c236ede3814ddb1.

This can lead to confusing situation where an aborted job due to
timeout is cleaned up by the "dev_mjeanson_system_clean" jobs before a
dev was able to have access the job logs. The current threshold for
cleaning up aborted jobs is 2 hours.

The "abort" notion is also used by the gerrit plugin to "abort" ongoing
build for which a new patch is set.

The most simple way to deal with this ambiguity is to revert this patch
and consider a timeout as a proper failure and let the human figure it
out.

A later patch will also add a timeout job description and update all
post-build actions depending on the aborted status.