jjb: Add "Depends-on: userspace-rcu:" for lttng-tools
While there we can remove some conditional actions based on `conf`
"static" and "debug_rcu", since they are not defined for the gerrit job. It
facilitate the condition handling for `depends-on` handling.

Babeltrace is now always pulled and does not depends on any "$conf"
conditional action.

Since that for the remaining `conf` ("std", "agents", "no-ust") userspace_rcu is
required, use conditional only to handle the `Depends-on` feature.

Signed-off-by: Jonathan Rajotte <jonathan.rajotte-julien@efficios.com>