jjb: binutils-gdb: filter out some known failures in GDB testsuite
The GDB testsuite is unfortunately known for not being clean.  In order
to get a somewhat useful CI job, filter out all the failures that can be
seen today, so that we can hopefully see the job passing.  With this,
any newly introduced failure will be visible, and not drowned in
pre-existing failures.

The ideal plan would be to work on fixing the tests listed here and
remove them from that list little by little.