Depends-on: take into consideration the git branch for query and default
Since gerrit cherry-picks have the same change-id the current query
returns multiples objects which is not supported. To alleviate this, we
use the GERRIT_BRANCH to refine the query. This is only valid for
project with the same version/name scheme at the branch level
(lttng-ust, lttng-modules). For now we need to handle this mostly for
those projects, it is quite rare that we need to handle this for
babeltrace and userspace-rcu.

Signed-off-by: Jonathan Rajotte <jonathan.rajotte-julien@efficios.com>
Change-Id: I5171d89083cbbb94e168158a11c69edebd4a766e