jjb: binutils-gdb: add target_board axis
Add a "target_board" testing axis.

"target board" is a DejaGNU concept that the GDB testsuite uses.  It
defines what kind of target system we are testing against.  In the case
of GDB, it defines through which means GDB debugs.  The default (when
nothing is specified) is the "unix" board, where GDB debugs against the
native system.  The GDB testsuite ships with some boards that can be
used as target boards with:

  make check RUNTESTFLAGS="--target_board=foo"

where foo is a name from gdb/testsuite/boards.  This patch adds an axis
to test each commit using the following target boards:

  - unix
  - native-gdbserver
  - native-extended-gdbserver

Unfortunately, the three axis values are completely independent, meaning
we build identical GDB three times.  In the future, if we could build it
once and run the testsuite multiple times, it would save some cycles.

Since the failures are different per target board, builds.sh defines one
"known-failures" file per target board.

Change-Id: I9531142b772dc4082d9782262d4daa18b166ac10