Ansible: add playbook to delete the .lttng folder of the jenkins user
The lttng-tools jobs currently do not use a custom LTTNG_HOME thus
lttng-sessiond uses /home/jenkins/.lttng as its rundir. Files in this
directory might not be always cleanup-up by lttng-sessiond during
testing. Normally this is not a problem but it is if someone (me) is
playing with the permissions of those files (lttng-sessiond side) and
end up setting a read only permissions for the user. This prevent
subsequent jobs from opening and overwriting the value of the files if
they are already present.

Removing the directory solves the problem for the CI infra.

Signed-off-by: Jonathan Rajotte <jonathan.rajotte-julien@efficios.com>
Change-Id: I102cd46368654999bd3d814dd0e09180700a129f