jjb: binutils-gdb: catch ERRORs in gdb.sum
When tests encounter a fatal error, it is reported as:

    ERROR: blah blah blah

Unlike normal test result lines, they don't include the name of the .exp
file currently running.  It would still be nice to catch them and
consider them failures, as they are abnormal.  Update the Python gdb.sum
processing script to remember the name of the current .exp, by looking
at the "Running" lines.  And when it encounters an ERROR, generate a
junit failure.

Change-Id: If99d73992241370afe9a9b59cb9abe491272d05f