lava: Rename inline x86-env-setup test
I propose renaming this job for two principal reasons: the setup is
run on more than x86 devices (it is used by kvm/qemu hosts as well),
and the purpose of this "test" is to create the vlttng virtualenv.

The more specific name also helps to avoid potential confusion with
the "x86-env-setup" that is run just beforehand.

Change-Id: I39915fec661f0519b94f68cbd6a26ca32ca3c254