lava: Switch from vmdebootstrap to debootstrap
vmdebootstrap is [abandoned][3] by it's author, who recommends no longer
using it. The package is no longer available in recent distributions.

While there is a more or less direct sucessors, [vmdb2][1] and other
options including [debos][2], the needs for creating the rootfs appear
to be minimal so I adjusted the script so any further modifications
are done by commands chrooted in the new environment.

[1]: https://vmdb2-manual.liw.fi/
[2]: https://github.com/go-debos/debos
[3]: https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=907751 "Request
for removal from Debian"

Change-Id: I0a9e5fa312e62dbce88dc1908433d7059ec294f1

lava: Fix value in generated rootfs' hostname

Change-Id: I748170655a04cb95a07713e8487e427e3ae536a3