jjb/lava: Parameterize repo URL and branch for test scripts
This permets us to launch a build with a different repo URL and/or
branch. Without this change, any changes we want to test must be
pushed on to the lttng-ci master branch before the run.

Pass repo URL and branch through to the clone in the setup job as well.

Change-Id: Ie9d45b2285e8e484dd23727278bb87e9c3ddfae2