bt benchmark: Collect commits to benchmark into a set
By collecting the the commits in to a set, we can give an indication
in the test runner how many jobs are going to be generated and
the overall progress of the lava job submissions.

The same commit hash may be present in multiple branches, using a set
prevents double checking the same hash twice.

Change-Id: I38a297f4147e4055d40cc462d0fdd7b10df88f46