lava: Remove redundant destructive tests
The majority of destructive tests are already being run in make check
via run-test-suites.sh. The only case not covered by that script was
the tests in 'root_destructive_tests' in the stable-2.12 and
stable-2.13 branches.

By adding support for that in run-tests-suites.sh, the
destructive-tests.yml can be completely removed from LAVA.

Change-Id: Ied4aceafb222cfe54c20ebff83608f702c1221e6