ansible: Deploy jenkin's SSH public key to standalone nodes
If ansible is run against any standalone nodes, the public set in
/root/.ssh/authorized_keys will be removed, and new connections will
fail until the VM is reset from the base-configuration snapshot if it
exists.

When deploying new nodes, ensuring the key is in the root user's
authorized_keys file in the first place avoids creating snapshots that
aren't accessible by the jenkin's user.

Change-Id: I37cc8e804aa4dea649df1558ccf8290d5df6be6f