jjb: Build & deploy lttng-www using jammy rootnode
For compatibility reasons, the job was designed to run against the
bionic node. It recently started failing to build since the
sass-embedded version was pinned to 1.53, and various artifacts
related to building it could no longer be fetched from maven.org.

In the process of unpinning it, changes were done so that the project
can be built on jammy.

Change-Id: I9084939fc2222dc6d0b8229790d9968127975ec3