ansible: Remove uuid set for ci-rootnode-deb12-i386-01-01
When the UUID is set, each time ansible is run the
community.libvirt.virt task will run and attempt to update the XML
definition of the virtual machine.

However, each time this happens the MAC address is changed. This
causes the VM to no longer get the expected IP address via DHCP.

If a new release of community.libvirt.virt sees the light of day, the
mutate option to keep the MAC address for interfaces could be tested
to see if it solves the problem we're seeing.

C.f https://github.com/ansible-collections/community.libvirt/commit/c4fe15801de0a1b61456acc8d2a5a704b0fe6ea5

For the moment, no changes to the XML template or settings are
expected for the VM, so removing the UUID is the simplest way to avoid
this issue.

Change-Id: I290999111da0eb29ccc856e55cd348da35e48df2