jjb: Derive lttng-modules architecture from platform label
This change updates the URL for fetching built kernels to key them by
build platform. This is done so that it's possible to differentiate a
kernel built on on deb12-amd64 (which will use gcc-12 by default) and
those built on bionic-amd64 which use gcc-7 by default.

An effect of this change is that all the kernels will need to be
rebuilt once the jobs are updated. It may be worth flushing the object
store of the linux kernels as this change is made so files in the old
paths are not kept endlessly without reason.

Change-Id: I96e0e6140827c660012e1264d96a324d8cca8801