ansible/jjb: Run lttng-tools 2.12 rootbuild on nodes with linux 5.x
The Debian 12 nodes run with Linux 6.1, which is not supported by
lttng-modules stable-2.12. Since the lttng-tools rootbuild jobs build
and test the modules component, those jobs failed.

A new libvirt instance was provisioned and both linux-image-amd64 and
linux-headers-amd64 installed from Debian bullseye and set to held so
there an instance with an appropriately versioned kernel to exercise
the lttng-tools rootbuild with lttng-modules tests.

Change-Id: Ie9234880b2785c5fa8d03dd7297fcc0d926b183c