lava: Install vlttng inside a venv
A number of the lava jobs were failing as the python tests couldn't
import the `bt2` module. While it may be possible to improve the
behaviour of vlttng to better address the situation two elements
combine to make the process work more systematically:

* Install `vlttng` inside a venv instead of using the
`--break-system-packages` option: once activated, this will set an
initial value for the `PYTHONPATH` environment variable
* When `vlttng` is invoked, since `PYTHONPATH` is no longer empty, the
value will be updated correctly.

In the case of the new rootfs, it also appears to be necessary to
ensure that `setuptools` is using `distutils` from Python stdlib,
although the module is deprecated and will be removed in Python 3.12

C.f. https://peps.python.org/pep-0632/

Signed-off-by: Kienan Stewart <kstewart@efficios.com>
Change-Id: If0aefb72f5c37721f88b49bd4a841b5ce61f46a0