ansible: Install gcc5 on Debian 12 CI nodes
lttng-modules crossbuilds for arm64 were failing for kernel versions <
4.4, principally due to gcc < 5.1 being explicitly denied in
lttng-modules.

To recover compilation coverage for these versions on this
architecture, gcc 5.5.0 has been built and deployed via ansible on to
the deb12-amd64 nodes.

The compilation coverage is not quite as large, as support for gcc 5
was only introduced in linux 3.18.

Fixes: support #1604
Change-Id: I49e7fc9d081c0b86cb25487d161efc3d1cf8f484
Signed-off-by: Kienan Stewart <kstewart@efficios.com>