lava: Remove apt update & upgrade from lava setup
The `apt-get upgrade` will block waiting for user input unless `-y` is
specified, causing builds to time out when there are actually updates
to perform.

Rather than forcing the command to run in non-interactive mode, I
decided to remove the update and ugprade. My principal reasoning is
that most of the jobs no longer require extra packages to be installed
at run time - they are included instead in the rootfs.

Change-Id: Ia9d76aa6ddda9b1a06b27125dca10f92b0e8c2fa
Signed-off-by: Kienan Stewart <kstewart@efficios.com>