ansible: Remove installation of curl from bookworm-backports
The configuration adjustment should be sufficient to avoid the
problem. It's not a big issue to use curl from bookworm-backports
otherwise, but the 'apt: upgrade' statement means that all the package
updates get run, violating an expectation for the default playbook
that only new packages will be installed - no updates should happen.

Change-Id: I8c083aa4d6a5d0a61a3924f52f295eab15467354
Signed-off-by: Kienan Stewart <kstewart@efficios.com>