jjb: integration: rework lttng-tools 32-64
Now that we have a libvirt plugin that allows using rootnodes with
snapshot reverts in pipelines steps, rework the 32-64 pipeline to remove
the external job and do every steps in the pipeline itself.

Rework the syntax for better readability, add the collection of
'test-suite.log' for easier tracking of test failure.

Change-Id: I87c4e10349bf86c71caaf4305c3b45a434aa5962
Signed-off-by: Michael Jeanson <mjeanson@efficios.com>