jjb: yocto23: replaced powerpcspe with powerpc
The Yocto 2.3 builder was moved from a p2020rdb to a p3041ds, this means
the CPU cores changed from e500v2 to e500mc. Both are 32bit big-endian
PowerPC but the e500mc is a more classic powerpc architecture while the
e500v2 is a rarer powerpcspe variant which has been deprecated.

Change-Id: If420a16cf97823c95908a6f0ebe9b34dc154d833
Signed-off-by: Michael Jeanson <mjeanson@efficios.com>