pyproject.toml: make lttng-ci a Poetry project
Run `poetry init` to put the boilerplate in pyproject.toml in order to
make lttng-ci a poetry project.

Use `package-mode = false` to indicate we don't intend to package this
project, but just use Poetry as a dependency manager.

Run `poetry add jenkins-job-builder` to add jenkins-job-builder as a
dependency.

Add some basic instructions to `README.md` on how to use poetry to
install and run the dependencies.

Change-Id: Ie0680d4c8e02ebd7120f80d3d28226034ef3142f
Signed-off-by: Simon Marchi <simon.marchi@efficios.com>