jjb: Remove daily trigger from binutils-gdb pipeline
The trigger is in the job definition in binutils-gdb.yml, and that is
more 'consistent'. A pipeline trigger isn't 'known' until the first
run, so a job must be run after update in order for it to be taken
into consideration.

This also avoids the `dev_review_` job starting on upstream changes.

Change-Id: I8798ac33b902ea6c1963a76382762cc256b66321
Signed-off-by: Kienan Stewart <kstewart@efficios.com>